Fix broken "Recently Added" XBMC images (Gotham)
The HTTP request to XBMC is returning a "Not Found" page, which is being
cached and served as an image by Maraschino. It appears Gotham doesn't
require the additional quoting.

I assume this will break Frodo, so it can only be used as an example as
to what works under Gotham.